Move KC distance testing into test_distance_matrics
